Refactor encrypt/decrypt methods to a base class
