+++ core/src/main/java/org/bitcoinj/crypto/BLSKeyExchangeCrypter.java
@@ -1,6 +1,5 @@
 /*
- * Copyright 2013 Jim Burton.
- * Copyright 2014 Andreas Schildbach
+ * Copyright 2019 Dash Core Group
  *
  * Licensed under the MIT license (the "License");
  * you may not use this file except in compliance with the License.
@@ -49,35 +48,9 @@ import static com.google.common.base.Preconditions.checkNotNull;
  * <p>2) Using the AES Key generated above, you then can encrypt and decrypt any bytes using
  * the AES symmetric cipher. Eight bytes of salt is used to prevent dictionary attacks.</p>
  */
-public class BLSKeyExchangeCrypter implements KeyCrypter {
+public class BLSKeyExchangeCrypter extends KeyCrypterAESCBC {
 
-    private static final Logger log = LoggerFactory.getLogger(KeyCrypterScrypt.class);
-
-    /**
-     * Key length in bytes.
-     */
-    public static final int KEY_LENGTH = 32; // = 256 bits.
-
-    /**
-     * The size of an AES block in bytes.
-     * This is also the length of the initialisation vector.
-     */
-    public static final int BLOCK_LENGTH = 16;  // = 128 bits.
-
-    /**
-     * The length of the salt used.
-     */
-    public static final int SALT_LENGTH = 16;
-
-    static {
-        // Init proper random number generator, as some old Android installations have bugs that make it unsecure.
-        if (Utils.isAndroidRuntime())
-            new LinuxSecureRandom();
-
-        secureRandom = new SecureRandom();
-    }
-
-    private static final SecureRandom secureRandom;
+    private static final Logger log = LoggerFactory.getLogger(BLSKeyExchangeCrypter.class);
 
     /**
      * Generate AES key.
@@ -103,85 +76,8 @@ public class BLSKeyExchangeCrypter implements KeyCrypter {
             log.info("Deriving key took {} for a BLS Key Exchange", watch);
             return new KeyParameter(key);
         } catch (Exception e) {
-            throw new KeyCrypterException("Could not generate key from password and salt.", e);
+            throw new KeyCrypterException("Could not generate key from BLS private and public keys.", e);
         }
-    }
-    /**
-     * Password based encryption using AES - CBC 256 bits.
-     */
-    @Override
-    public EncryptedData encrypt(byte[] plainBytes, KeyParameter aesKey) throws KeyCrypterException {
-        checkNotNull(plainBytes);
-        checkNotNull(aesKey);
-
-        try {
-            // Generate iv - each encryption call has a different iv.
-            byte[] iv = new byte[BLOCK_LENGTH];
-            secureRandom.nextBytes(iv);
-
-            ParametersWithIV keyWithIv = new ParametersWithIV(aesKey, iv);
-
-            // Encrypt using AES.
-            BufferedBlockCipher cipher = new PaddedBufferedBlockCipher(new CBCBlockCipher(new AESEngine()));
-            cipher.init(true, keyWithIv);
-            byte[] encryptedBytes = new byte[cipher.getOutputSize(plainBytes.length)];
-            final int length1 = cipher.processBytes(plainBytes, 0, plainBytes.length, encryptedBytes, 0);
-            final int length2 = cipher.doFinal(encryptedBytes, length1);
-
-            return new EncryptedData(iv, Arrays.copyOf(encryptedBytes, length1 + length2));
-        } catch (Exception e) {
-            throw new KeyCrypterException("Could not encrypt bytes.", e);
-        }
-    }
-
-    /**
-     * Decrypt bytes previously encrypted with this class.
-     *
-     * @param dataToDecrypt    The data to decrypt
-     * @param aesKey           The AES key to use for decryption
-     * @return                 The decrypted bytes
-     * @throws KeyCrypterException if bytes could not be decrypted
-     */
-    @Override
-    public byte[] decrypt(EncryptedData dataToDecrypt, KeyParameter aesKey) throws KeyCrypterException {
-        checkNotNull(dataToDecrypt);
-        checkNotNull(aesKey);
-
-        try {
-            ParametersWithIV keyWithIv = new ParametersWithIV(new KeyParameter(aesKey.getKey()), dataToDecrypt.initialisationVector);
-
-            // Decrypt the message.
-            BufferedBlockCipher cipher = new PaddedBufferedBlockCipher(new CBCBlockCipher(new AESEngine()));
-            cipher.init(false, keyWithIv);
-
-            byte[] cipherBytes = dataToDecrypt.encryptedBytes;
-            byte[] decryptedBytes = new byte[cipher.getOutputSize(cipherBytes.length)];
-            final int length1 = cipher.processBytes(cipherBytes, 0, cipherBytes.length, decryptedBytes, 0);
-            final int length2 = cipher.doFinal(decryptedBytes, length1);
-
-            return Arrays.copyOf(decryptedBytes, length1 + length2);
-        } catch (InvalidCipherTextException e) {
-            throw new KeyCrypterException.InvalidCipherText("Could not decrypt bytes", e);
-        } catch (RuntimeException e) {
-            throw new KeyCrypterException("Could not decrypt bytes", e);
-        }
-    }
-
-    /**
-     * Convert a CharSequence (which are UTF16) into a byte array.
-     *
-     * Note: a String.getBytes() is not used to avoid creating a String of the password in the JVM.
-     */
-    private static byte[] convertToByteArray(CharSequence charSequence) {
-        checkNotNull(charSequence);
-
-        byte[] byteArray = new byte[charSequence.length() << 1];
-        for(int i = 0; i < charSequence.length(); i++) {
-            int bytePosition = i << 1;
-            byteArray[bytePosition] = (byte) ((charSequence.charAt(i)&0xFF00)>>8);
-            byteArray[bytePosition + 1] = (byte) (charSequence.charAt(i)&0x00FF);
-        }
-        return byteArray;
     }
 
     /**
